housekeeping/manager: Use '--auto' with 'git-pack-refs(1)'
The housekeeping manager, calls 'git-pack-refs(1)' with the '--all'
flag. For the files backend, it always repacks all loose refs into a
single packed-refs file. For the reftable backend, it always
autocompacts all reftable tables into a single table.

Let's instead use the newer '--auto' flag. This follows the same
heuristics as we have in Gitaly. As such, for the files backend, it
allows certain loose refs to packed-ref ratio. For the reftable backend,
it will modify the table to exist in a geometric progression. This could
help reduce the number of file writes during housekeeping.