chore: Remove toolchain directive from go.mod
The toolchain directive specifies the version of the Go toolchain which
should be used to build the package, regardless of the defined `go`
version. The feature was introduced to allow a package's compatibility
version to remain the same, but also take advantage of new compiler
improvements.

Since Gitaly is mostly compiled via externally-managed build images [1]
which include a specific Go toolchain, the `toolchain` directive is not
required and can result in the wrong compiler being downloaded and used.

The directive was introduced into Gitaly at some point by the Renovate
bot. An MR [2] has recently been merged to stop the bot from adding the
directive in the future.

Remove existing instances of the `toolchain` directive, and ensure a
semver value is provided for the `go` directive. This prevents
`go mod tidy` from re-adding the `toolchain` directive.

[1] https://gitlab.com/gitlab-org/gitaly/-/blob/master/.gitlab-ci.yml#L10
[2] https://gitlab.com/gitlab-org/frontend/renovate-gitlab-bot/-/merge_requests/1152