chore: Set GOTOOLCHAIN=local
Since we have removed the `toolchain` directives from the `go.mod`
files, Go will start treating the `go` directive in a similar manner.

For example, if we update the `go` directive to `1.24` and try to build
the project with a Go 1.23.9 toolchain, the toolchain will automatically
download and use Go 1.24.* transparently. This can happen in the test or
build process in CI.

By setting GOTOOLCHAIN=local, we can prevent this from happening. Rather
than downloading and running a newer, compatible toolchain, the build
simply fails with an error like below:

  go: go.mod requires go >= 1.24 (running go 1.23.9; GOTOOLCHAIN=local)

We want the build to fail because we should ensure the `go` directives
are always bumped in sync with the `go_supported` array in the CI
definition. This ensures we're testing all supported toolchains
properly.