test: Show advice for compilation issues
Some of our tests compile various Gitaly binaries as part of the test
setup. When the user has multiple Go toolchains installed on their
system that are managed with tools like asdf and mise, and execute tests
via an IDE that isn't aware of these tools, compilation can fail due to
version mismatches.

Intercept the error from the compiler and display some useful advice if
this happens.