housekeeping: Don't log LookupMethod errors
The housekeeping middleware intercepts all incoming gRPC requests, which
includes gRPC methods not registered with Gitaly's
protoregistry.Registry. An example is the Check() method of the Health
endpoint, which is an implicit gRPC server feature.

The middleware calls Registry.LookupMethod() and errors if an error is
returned. The only error that can possibly be returned is if the lookup
failed. If this happens, we should not log an error as it's likely due
to intercepting a healthcheck RPC. If it's another problem, we can defer
the handling to the next middleware in the stack.

We can't add the health service to the Registry because we expect it to
define [1] Gitaly-specific metadata like OpType.

[1] https://gitlab.com/gitlab-org/gitaly/blob/f14ce9130d4b339716df4f04ada14ca505c1464f/internal/grpc/protoregistry/registry.go#L50-L53