Test duplicate start points in TestWalkObjects
TestWalkObjects is asserting that the object walk produces correct
list of object IDs. It's expected that the walk does not produce
duplicates even if there are duplicate objects in the graph, or
we start the walk from duplicate start points. Add tests to assert
duplicates do not change the result.