Disable bundle URI in CICD
Custom runners were created to support bundle URI
in the gitaly CICD pipelines. They can be used
with the runner tag `bundle-uri`.

However, those runners do not exists on the
security project, wihch runs the job to scan
for security issues. This is causing the jobs
to be stuck because no runners are available.

This commit reverts the enabling of bundle URI
for the Gitaly project in order to unblock the
security pipeline. We will find a way to make it
work with bundle URI at another time.