linguist: Increase read limit to 16KiB
When we initially implemented [1] language detection in Go with go-enry,
we set an arbitrary read limit of 2048 bytes for each object being
processed. This allows us to prevent reading especially large objects
into memory, and typically a small slice of the file is enough to
determine the language used.

Recently we've seen an edge case [2] where a TSQL file was mistakenly
identified as a plain SQL file. Since plain SQL files are considered
"data" rather than "programming" by go-enry, the stats were discarded
and caused the language bar in the repository UI to render incorrectly.

We can increase the read limit to allow go-enry to process more of the
file. I computed the size distribution [2] of .go files in the Gitaly
repository to get a rough idea of sizes:

  1k:    560
  2k:    356
  4k:    331
  8k:    229
 16k:    114
 32k:     53
 64k:      9
128k:      5
256k:      1

This doesn't cover other languages of course, but gives us an idea of
where the new limit could sit. With a 16KiB limit, we can cover the vast
majority of files in this example. The memory and performance impact
should also be minimal; on `gprd` we see anywhere from 4-15 RPS over a
24-hour period on the CommitLanguages RPC which executes this code, and
it's a non-blocking call made by a Sidekiq worker.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4580
[2] https://gitlab.com/gitlab-org/gitlab/-/issues/508982
[3] https://superuser.com/a/1100340