Remove micro optimization for avoiding rev-list commands
packObjects method contains a micro optimization to avoid starting
Git commands if there are no objects to pack. The code used to
be somewhat different. It launched more commands, and some of them
would fail if there were no objects to pack.

These days we're launching a rev-list command first to walk the
quarantined objects, and conditionally runnning the other packing
commands if needed. This optimization mostly serves to avoid running
the rev-list command if there are no writes into the quarantine. It's
doubtful that this has much impact on performance given running a
single command on outside of the critical section isn't very
critical. Writes are also relatively rare, so these commands
wouldn't get spawned that often.

Remove the unnecessary optimization to simplify the code.