Fix flaky reftable migration tests
The reftable migration tests are not synchronized correctly which
leads to flakiness.

The test is accessing the repository directly on the disk to check
its reference backend. By this time, it's guaranteed the migrator
has committed the transaction. However, it isn't guaranteed the
transaction has been applied. The test may assert against an arbitrary
state in-between applying the transaction. This leads to test failures
when the repository state doesn't match what Git expects.

This has been patched around by running the test in a loop three
times. This may hide some failures but doesn't actually fix the
synchronization issue.

Fix this problem by checking the reference backend through the API
as normal clients would.