partition: Backup KV state
We previously included the KV state of a partition during backup,
but this state only contains the KV entries for tx.KV() which is
storing the keys under "kv/" prefix. But there are some other keys
needed for full restoration, such as applied_lsn and relative_path
which are not stored under same prefix.

In this MR, we are adding new method to transaction manager that
will return raw unprefixed KV data of the partition.

With this new structure, we will just need to add "p/{$partitionID}"
prefix to the all keys during restore as that is a common prefix for
applied_lsn and tx.KV() entries.