Fix reftable hash ID parsing
NewReftable parses a reftable's header and footer. V2 reftables contain
the hash ID used as the last header field. The current parsing code is
not correctly extracting the hash id from the footer which contains a
duplicate header. It's skipping past the entire header and trying to
read the hash ID afterwards. In reality, the hash ID is the last field
of the header.

Fix the parsing code to correctly read the hash ID from the footer.