offloading: Remove logic that modifies the alternates file
Previously, we attempted to add an entry to the alternates file to
allow downloading pack files to that location temporarily.
However, as discussed in
https://gitlab.com/gitlab-org/gitaly/-/issues/6779, this approach
violates the read-only principle of snapshot-based transactions.

This commit removes the logic that modifies the alternates file
during offloading to ensure alignment with read-only transaction
guarantees.