chore: Update Go to 1.23.9 in .tool-versions
The version of Go 1.23 used in the build images was updated to 1.23.9
over a month ago [1]. This effectively makes Go 1.23.9 the default
version of the Go toolchain that we use in CI jobs.

Update the .tool-versions file to match. This ensures parity in
toolchain versions we use locally and in CI.

[1] https://gitlab.com/gitlab-org/gitlab-build-images/-/blob/1961fceb869afa3067cced5d65cf919ea151a91b/scripts/lib/custom-docker-build#L81