storage: Add option to skip preventReftableCompaction
The preventReftableCompaction function creates a lock file in the
snapshot when reftable is enabled. This is a workaround to prevent
new tables written during a transaction from being merged
with existing ones
(see https://gitlab.com/gitlab-org/git/-/issues/350).

However, during leftover file migrations, we don’t want the migration
task to mistakenly identify this lock file as part of the
original repository and attempt to delete it, leading to an error
since the original repo does not contain this file.

This commit introduces an option to skip preventReftableCompaction.
This option is enabled when the transaction is created
specifically for a migration task.