Introduce a reftable package
We have some preliminary code for interfacing with reftables in the
Git package. We'll soon need more, so let's create a separate reftable
package where we can collect all reftable specific code. This way we
have a dedicated namespace for all reftable related code, and don't need
to prefix everything with reftable as the package name scopes the names
clearly on its own.