Streamline naming of reftable package members
Now that we've moved reftable specific logic into the reftable
package, the naming of several members stutter as they repeat
reftable in their name. Since reftable is now part of the type
name already through the package name, streamline the naming of
the members and drop the reftable references from them.