Determine header and footer size from the version on demand
We're currently storing the reftable's header and footer size on
the returned reftable struct. This seems unnecessary given these are
dependent only on the version of the reftable file. Simplify the type
by determining the values on demand from the version of the table we
are already storing in the footer.