Pass down the source file when parsing references
IterateRefs returns a list of reference in the reftable. This method
is only used in the tests and operates on the entire reftable file.
Due to this, NewReftable currently takes in the entire reftable file.
This is fine for testing but we soon want to start using parts of the
reftable parsing code in production, and we don't want to read the
entire file always. To make space for this, pass down the src through
the call stack of IterateRefs. This way we can take a reader in later,
and isolate reading the entire file only to the method that requires
that.