Parse reftable from an io.ReadSeeker
We're currently taking a byte slice containing the entire reftable
in NewReftable. This isn't optimal as it requires reading the entire
file to even parse out the information in the header and footer of
the file.

Take in an io.ReadSeeker instead. This way we can pass for example
a file descriptor pointing to a reftable file, and only read in
what we need to.

We still read the entire file in IterateRefs. This method is only
used in tests, so we don't need to optimize it yet.