Export Table type
We've exported methods that return the reftable type but we haven't
exporte the actual type. This is a bad practice as one can't publicly
access the type that is returned.

Rename the type to Table to reduce stuttering, and export the type
so it is accessible from where it can be used. In the same go, rename
NewReftable to ParseTable to better describe what it does.