Handle rollback when raft is toggled on/off
This commit handles the scenario when raft was enabled at one stage,
migration has completed and then it was disabled after that. At this
point, the config for raft indicates it's disabled yet the db for a
particular storage will indicate the migration has already been done.
We will rollback and revert the migration in this instance so that
the partition structure will follow the original structure without
support for raft replicas.