backup: Refactor ResetRefs function
The ResetRefs function restores a repository's references to match
the refs state from a backup by deleting existing references that
are not present in the backup set, and creating or updating other
references from the backup list. Currently, it handles all updates
within a single request, which works well in most cases. However,
we aim to enable optimistic updates in certain scenarios, allowing
the overall restore process to proceed even if some references fail
to update. To achieve this, we will separate removal updates from
other updates (such as creating new reference). Removal operations
are unlikely to fail since we are confident that the reference exists
prior to deletion. A potential issue arises when updating a reference
to point to an object that does not exist in the repository.
By distinguishing these two update processes, we can allow the regular
ref updates to fail without affecting the removal updates.

This change is a preperation to get rid of the recreateRepo logic
during restoreFromBundle, and will be utilized in the following commit.