backup: Refactor restore from bundle
Currently, if ResetRefs returns an error, we fall back to restoring
from a bundle. ResetRefs can fail if some references point to objects
that no longer exist in the repository. When restoring from a bundle,
we first recreate the repository before fetching the bundle, because
FetchBundle does not remove references that are already present in the
repository but missing from the bundle. This flow normally works fine
but it can lead to conflicts when the reftable backend is enabled, as
the reference recording logic during the fetch bundle process performs
conflict check on the original state of the repository rather than the
recreated version.
Therefore, instead of recreating the repository prior to restoring the
bundle, we will first restore the bundle and then perform a reference
reset on the restored state. Since we previously refactored the way
ResetRefs works, it will perform ref removal updates separately. And
this will get rid of the refs exists in the repository but not in the
bundle scenario. We will also still perform regular ref updates but
we will do it viaa the `optimistic` parameter, which will simply ignore
any errors.