raft: Use event registry for ProposeConfChange
In membership management methods we use the ProposeConfChange interface
by etcd/raft, which is asynchronous. This can cause a race between the
caller of membership methods and Raft's replica state machine, in
situations where caller receives incorrect info regarding the leader or
the contents in routing table. So, to track the entry, we need to use
the ReplicaEventRegistry.