raft: Make PartitionKey opaque
The composite PartitionKey of (storage_name, partition_id) didn't make
sense once the partition crossed the node boundary and caused a bit of
confusion. The intention of the key is to remain immutable after  being
minted by the node that first created the partition. Other nodes that
are replicating the partition were expected to use the same PartitionKey
for lookups, but maintain their own internal storage and partition ID
mapping.

Make the key opaque so there's a clear distinction between the immutable
key minted by the authority node, and the storage and partition ID
attributes used by the local nodes to store the replicated partition.
Several changes were also required:

Modify the RoutingTable's ApplyReplicaConfChange() function to take an
explicit storageName parameter, representing the name of the storage
that will be storing the replicated partition. The routingKey is also
modified to use the partitionKey directly.

Modify the SendMessage handler to validate only that the value of the
key exists.

Modify the filename generated by the SendSnapshot RPC to use the full
PartitionKey value.