raft: Rename authorityName to storageName
The ReplicaLogStore contains an authorityName field, and the
NewReplicaLogStore() constructor takes an authorityName parameter. This
name only makes sense if the ReplicaLogStore is being created by the
storage which is minting the partition. When the store is being created
on another node due to the partition being replicated, the argument
would be the local storage name.

Rename authorityName to storageName to accurately represent what is
being passed in.