fix: Use binary partition ID consistently
The NewKVRoutingTable() constructor creates a prefixed Badger database
within the dedicated metadata partition, with partition ID = 1. The
partition is reserved for Raft use, so normal repository partitions
begin at ID = 2.

Partition prefixes in Badger are of the form p/<partition_id>/...
where partition_id is assumed to be encoded as binary with a fixed
length. Additional keys relative to the partition can be nested
underneath the prefix (the ... portion).

The partition manager's ListPartitions() function uses a
partitionIterator to attempt to retrieve a list of active partitions
based on the available keys in Badger. Since there can be many keys that
all begin with the same p/<partition_id> prefix, the
extractPartitionID() function [1] is responsible for extracting only the
<partition_id> portion of the key. It does this by assuming the length
of the binary-encoded <partition_id> and unmarshals it into an integer.

Because the KVRoutingTable doesn't encode the partition ID as binary, the
extractPartitionID() function will unmarshal past the end of the
partition ID part of the key, returning a nonexistent partition ID. This
also has the side effect of breaking one of the partition backup tests
[2] as it changes the number of expected partitions. Currently, multiple
instances of the invalid prefix would unmarshal to the same invalid
partition ID due to the beginning of the keys being the same. For
example:

- p/1/raft/default/abc and
- p/1/raft/default/def

would produce the same invalid partition ID because:

- We start with the key p/1/raft/default/2
- p/ is stripped out, leaving 1/raft/default/2
- We take the first 8 bytes of the key, as we assume the PartitionID's
  64-bit integer occupies the entire space, leaving 1/raft/d.
- 1/raft/d is unmarshaled into an invalid partition ID of
  3544177194420154212.

When we transition to using a SHA256 hash as the partition ID, the
beginning of the hash would be different, causing an additional partition
to be detected incorrectly and returned.

Fix this by exporting the storagemgr.KeyPrefixPartition() function and
using it in the NewKVRoutingTable constructor.

[1] https://gitlab.com/gitlab-org/gitaly/-/blob/9a72666ea3a0a18d2f02d2821b880daee41d9767/internal/gitaly/storage/storagemgr/partition_manager.go#L763-777
[2] https://gitlab.com/gitlab-org/gitaly/blob/b52f1ecc47e28d380c62a01deedcff7abedc86fd/internal/backup/partition_backup_test.go#L46-50