git: Modify tests to be ready for batched updates
The 'next' branch of Git includes the patches to use batched reference
updates. Traditionally reference updates in Git, specifically for the
'git-fetch(1)' and 'git-receive-pack(1)' commands, are performed one at
a time.

This is not so-performant, the batched reference update mechanism solves
this by batching all the reference updates in a single batch. Within
which individual updates can pass/fail independently. Since reference
updates are batched, the number of calls made to the
reference-transaction hook change with this series. Accomodate the same
in Gitaly.