Rename IterateRefs to GetReferences
IterateRefs is returning a slice containing all references within
a reftable file as opposed to returning an actual iterator. While
working with an iterator would be better to avoid loading all
references into a slice, it requires more work to adapt the existing
code. Rename the method for now with a more appropriate name.