Fix edge case when limit is not provided
**Problem**

Changes from https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7979
caused the Rails tests to fail. The root cause was an edge case when
page token is set but limit is not provided.

Before, this combination would raise an error. But
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7979 changed that
and allowed to proceed with the execution.

**Solution**

Return an error when Limit is set to zero to match the previous
behavior.

Changelog: fixed