ref: Add RefIterator
The current pattern of getting a branch with its commit is to call
for-each-ref, and then call cat-file separately. However, this is
unnecessary as the --format of git for-each-ref can be used to gather
all the necessary commit information.

Create a iterator that a caller can use to iterate through a git
for-each-ref result. The fields are delimited by a null \x00 character
and the lines are delimited by a \n\x00 combination delimiter because
there may be \n characters in the content body.