Record reftables directly from the transaction snapshot
TransactionManager currently records reference writes done with the
reftable backend by replaying them on the latest state of the repository
using 'git update-ref' and recording the created table files. Replaying
the changes on the latest state was a simplistic way to capture the
updates but it leads to bad performance. The problem is that we're using
Git to do the replay, and Git requires a repository snapshot to replay
the operations in. This is prohibitively expensive to do in the critical
section, and would lead to performance issues if reftables were to be used
in a production environment.

Files backend doesn't have this issue as we capture the reference updates
as they were written into the snapshot during the transaction. This way we
don't need to create repository snapshot to replay the operations in. This
optimization hasn't so far been applied to reftables.

This commit applies the same optimization to the reftable backend, and
starts capturing the table files as they were written out during the
transaction. This way we don't need to replay the operations and the
associated snapshot.

Reftables unfortunately produce conflicts on every single concurrent write.
The tables contain an update index which is roughly a sequence number for
the writes. Update index for a new write is always one higher than the
previous one. As the writes happen in snapshots, concurrent Git operations
will assign the same update index for their tables. Similarly each write
appends the created table to the tables.list file which would lead to each
transcation producing a conflicting tables.list file.

TransactionManager thus needs to resolve these conflicts at commit time.
This is done by resequencing the tables before they are logged. We take a
look at the latest state in the repository, and patch the update indexes and
the tables.list file to push the new tables at the end of the sequence.
this way concurrently writes that produced conflicting update indexes get
resequenced one after another, and can be committed into the repository
without having to rewrite the entire table.

Since we're now recording the tables from the snapshot, the replaying logic
is removed.