Implement logic for patching reftable update indexes
If multiple transactions concurrently update references with transactions,
they won't see the modifications each of them makes. This causes them to
allocate conflicting update indexes when they are writing reftables. This
would require transaction manager to abort one of the transactions, meaning
that we couldn't process any concurrent reference updates with reftables.

To avoid this, we'll instead resolve the conflicts, and resequence the table
files written out by the transactions. To do this, TransactionManager will
patch the update indexes in the reftables written out by transactions.

This commit implements a method for patching the update indexes in a
reftable file. The min and max update indexes are stored in the header and
the footer of the reftable file. The footer additionally contains a checksum
that covers the update indexes in the footer, so we also update the checksum
to account for the patched values.