Handle file management in Table type
We're currently taking an io.ReadSeeker into ParseTable. Take a file
path instead and handle the file management internally within the type.
This allows us to use methods such as ReadAt on the file type to do
reads without first performing a seek. Similarly, using a file directly
will help us later with patching the update indexes as we can write
into the files using WriteAt, and sync the file.