Add a general test helper for switching out assertions
We have helpers for switching out assertions based on conditions like
whether or not raft or WAL is enabled. Add a general function to do
this without checking a specific condition. This way the function
is more reusable so we can use it also in scenarios where we need
to switch an assertion but it doesn't make sense to add yet another
helper for a specific condition.

Replace the implementations of the existing helpers with a call to
the new general helper.