raft: Fix replica registration in AddNode test case
Earlier we were not registering the replica properly due to which the
messages were not being stepped in the destination node, the transport
instance used in the replica and the mock server are different. As a
result, we need to register the replica separately in the transport
associated with the mock server. Several other related issues were also
identified - https://gitlab.com/gitlab-org/gitaly/-/issues/6819, they
will be addressed in subsequent merge requests.