reftable: Register migrations at the end of MUTATOR RPCs
The reftable migrator only registers a migration when a ACCESSOR RPC
comes in and registers a cancellation for all other types of RPCs. This
is to ensure we don't create conflicts.

This however has a week point. Write heavy repositories will always keep
registering cancellations whenever a write comes in. And if only writes
are being made to the repository, it wouldn't even register a migration.
By initiating a migration at the end of a write cycle, we can avoid
conflict while also ensure that migrations are registered even if the
repository only sees writes.