diff: Add diff filter to FindChangedPaths RPC
The FindChangedPaths RPC performs a git-diff-tree(1) to return raw diff
info. This Git command also has the `--diff-filter` flag which filters
the output to only contain diffs for the specified status types. In the
FindChangedPaths RPC, the `--diff-filter` flag is hardcoded to "AMDTCR".
Some RPC client may want to have a subset of these diffs returned
though.

Introduce the diff_filter field to the FindChangedPaths RPC request
which allows specific status types to be targeted. When left
unspecified, the default is to use the previously hardcoded value.