commit: Use git-last-modified(1) for ListLastCommitsForTree RPC
At the moment we do N+1 Git calls to fill in the
ListLastCommitsForTreeResponse. To feed this data, we first do `git
ls-tree` to get the list of tree entries, than get the correct page, and
run `git log --max-count=1` for each path. This is ineffecient.

We've been working hard to upstream the git-last-modified(1) command to
Git[1]. This command can be used to replace all the git-log(1) calls.
Implement it's use behind a feature flag.

[1]: https://gitlab.com/gitlab-org/git/-/issues/114