reftable: Fix bug where cancellation func wasn't set
The reftable migrator stores a cancellation func in the current state,
so any incoming writes to the repository can trigger a cancellation
which would stop the migration to avoid conflicts.

However, there is a small bug here where if there was already existing
state, we don't store the cancellation. This is because we use a
`LoadOrStore`, but don't store if the Load branch is triggered. Fix this
by explicitly storing in such a situation. Add a test for the same.