raft: Use PartitionKey in Replica constructors
The RaftReplicaFactory function and the NewReplica constructor currently
take the storageName and partitionID as arguments.

The storageName will continue serving a purpose as the name of the
storage on the newly-added node where we want the replicated partition
to reside. The partitionID however will no longer be needed. When a
partition is replicated to a new node, the new node's partition manager
will take care of assigning a new partition ID. We threaded the ID as
an argument previously as a way to compose the PartitionKey, so this can
now be removed.

Replace partitionID with PartitionKey, and add some TODO breadcrumbs for
important follow-ups.