raft: Rename PartitionKey to RaftPartitionKey
Adding the Raft prefix reduces ambiguity about where and how the key
should be used.

Other areas of the codebase define similar sounding keys, such as the
backup.PartitionInfo composite key, and the storage.PartitionID which
represents a locally-allocated partition ID.

We should make it clear that the RaftPartitionKey is meant to represent
a key that is unique across an entire Raft cluster, and can be provided
to any node of that cluster in order to retrieve its local replica of
a partition.