reftable: Increase the reftable test coverage
Our CI systems run jobs independently for the WAL and for the reftable
backend. But we don't have one which runs for both of them together.
Let's increase our test coverage for reftable by adding a matrix to run
reftables with praefect, wal, raft and praefect-wal.

This includes some fixes to make sure the pipeline works as expected:

1. In 'reftable/middleware_test.go' we used to return early if the
default reference backend was reftables. This causes a goroutine leak
since we don't close the reftable migrator when returning early. While
one fix would be to also close the migrator while returning early. Let's
instead let the test run its course, this way, we also validate that the
migrator works as expected when working on a already migrated
repository.

2. In `TestCalculateChecksum`, it rewrites the reftables file by
replacing the HEAD ref with an invalid one. When the WAL is also
enabled, the file becomes read-only as it belongs to the canonical
repository. To cicumvent the read-only permissions, delete the file
first before calling os.WriteFile.

3. Finally, skip running `TestRestoreRepository` with reftables for now.
This needs fixes in the `backup` package which is being handled in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7971/.