test: Rerun failed tests in CI only
In fb75a2f11 (test: Enable the --rerun-fails flag, 2025-03-13) we added
the use of the --rerun-fails flag for `gotestsum`, which will re-run
failed tests to reduce the burden of dealing with flaky tests.

This flag was specified unconditionally, which can be annoying if you
are trying to run tests locally which _should_ fail.

Limit the use of --rerun-fails to CI only, by checking the `CI`
predefined variable [1]. Running the `make test` target locally will no
longer cause failed tests to be re-run.

[1] https://docs.gitlab.com/ci/variables/predefined_variables/