streamcache: Refactor the streamcache to make it reusable
Currently, the streamcache is tightly coupled to the packobjects
cache. As an example, when creating an instance of the cache, a
metric is used that bears the name `gitaly_pack_objects_cache_enabled`.

This is problematic, because it makes the cache not reusable for
other purposes than packobjects.

This commit refactors the cache
to make more reusable. It adds a new `name` field in the streamcache
configuration, and uses that name to create the metrics. This ensure
that we don't break any current metrics for the packobjects cache
and that we don't break any dashboard. At the same time, it allows
to reuse the cache for other purposes.