raft: Implement JoinCluster RPC
This commit implements the JoinCluster RPC, which currently updates the
routing table using the peer list received from the leader. Replica
bootstrapping is not included in this commit and will be addressed in a
separate MR. End-to-end tests for `AddNode` are also not included at
this stage, as the raftmgr package mocks the JoinCluster RPC, rendering
the test ineffective. The tests  will be added in the service/raft
package as part of the bootstrap MR.