repository: Use correct backend in `ReplicateRepository`
The `ReplicateRepository` RPC is used to replicate an existing
repository into a target repository. To do this, it extracts a tar of
the existing repository then carries over the config.

So if the existing repository is using the files backend, while the new
repository is created using reftables (because the reftable featureflag
was turned on), the new repository will end up with a 'reftable' folder.
But since we copy over the config after the tar extraction, the repo
gets converted to a files backend repository.

This is weird and wrong and leaves stale files/folders behind. Let's
explicitly set the reference backend of the new repository to be the
same as that of the source repository. This would avoid any such issues.