repository: Migrate newly created repositories via ReplicateRepository
In the previous commit, newly created repositories via the
`ReplicateRepository` RPC was modified to ensure that the reference
backend used is the same as that of the source repository.

There is also a featureflag to enable reftables for new repositories and
this would violate that understanding. So let's manually migrate the
repository if necessary.