git: Add GitMaster feature flag
This flag when enabled introduces a new execution environment which uses
the git's master version which is 7–14 days old.

The Git master version used in the Makefile reports as "2.50.git" but
contains the batched reference updates feature, while the mentioned
threshold is 2.52. So we check if the flag is enabled to determine if
batched updates are in use during tests. We couldn’t remove the check
entirely because our fallback version, the stable Git version 2.50, does
not include this feature.