benchmark: Add Terraform resources to create git-repositories disk
The existing documentation and Terraform plan reference a GCP disk image
called "git-repos", which contains a collection of cloned repos that are
used for benchmarking. The disk image appears to have been created
manually.

In order to simplify the process of recreating the disk image, and to
allow the list of repositories to be easily customised, add Terraform
resources to spin up a temporary GCP VM which clones a configurable list
of repositories onto a secondary disk. The secondary disk is turned into
a disk image which is later mounted onto the Gitaly host VM.