benchmark: Introduce Go tool for executing workloads
The existing benchmark-gitaly.sh script is a thin wrapper around the
`ghz` tool, allowing for a workload to be executed against a single RPC
and repository. We need the ability to generate more elaborate workloads
that target a mix of RPCs and repositories. A small tool written in Go
would be ideal, as it's a language familiar to the team and easily
allows us to run tasks concurrently.

Create a Go program that takes a `workload.yaml` file as input, which
defines a collection of requests to execute against Gitaly. Right now,
the workloads are executed sequentially and written out to the same
directory as before. A later commit will integrate this tool with the
Ansible tasks and remove the benchmark-gitaly.sh script.

Also create a sample workload.yaml file in the testdata/ directory with
two benchmarks, and create the associated FindCommit query file for the
gitlab.git repo.