storagemgr: Skip preventing reftable compaction for MigrateReferenceBackend
The transaction manager, skips reftable compaction by adding a lock file
on the latest reftable table. This is required to prevent autocompaction
during writes. Without which we would have conflicts.

But the MigrateReferenceBackend migrates between reference backends and
shouldn't be subjected to this. Since it deletes the entire 'reftable'
folder when migrating from reftable -> files backend. Let's add a map
which allows RPCs to skip this the compaction prevention.