benchmark: Introduce a config-workload.yml file
A previous commit introduced a Go tool for executing benchmarks, along
with a testdata/workload.yaml file serving as a sample configuration.

Hoist the configuration file to the root level so it's easy for the user
to customise. Rename the existing config.yml to
config-infrastructure.yml, as it was primarily responsible for
customising the Terraform deployment. Alongside this file, introduce a
config-workload.yml which feeds directly into the Go tool.